pkg/prometheus-cachethq/common/utils/server.go
package utils

import (
	"fmt"
	"net/http"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/oxyno-zeta/prometheus-cachethq/pkg/prometheus-cachethq/common/errors"
)

const forwardedLength = 2

func GetRequestURL(r *http.Request) string {
	scheme := "http"
	if r.TLS != nil {
		scheme = "https"
	}

	return fmt.Sprintf("%s://%s%s", scheme, RequestHost(r), r.URL.RequestURI())
}

func RequestHost(r *http.Request) string {
	// not standard, but most popular
	host := r.Header.Get("X-Forwarded-Host")
	if host != "" {
		return host
	}

	// RFC 7239
	host = r.Header.Get("Forwarded")
	_, _, host = parseForwarded(host)

	if host != "" {
		return host
	}

	// if all else fails fall back to request host
	host = r.Host

	return host
}

func parseForwarded(forwarded string) (addr, proto, host string) {
	if forwarded == "" {
		return
	}

	for _, forwardedPair := range strings.Split(forwarded, ";") {
		if tv := strings.SplitN(forwardedPair, "=", forwardedLength); len(tv) == forwardedLength {
			token, value := tv[0], tv[1]
			token = strings.TrimSpace(token)
			value = strings.TrimSpace(strings.Trim(value, `"`))

			switch strings.ToLower(token) {
			case "for":
				addr = value
			case "proto":
				proto = value
			case "host":
				host = value
			}
		}
	}

	return
}

func AnswerWithError(c *gin.Context, err error) {
	// Try to cast as common error
	// nolint: errorlint // Ignore this because the aim is to catch project error at first level
	err2, ok := err.(errors.Error)
	// Check if cast was a success
	if ok {
		c.AbortWithStatusJSON(err2.StatusCode(), gin.H{
			"error":      err2.Error(),
			"extensions": err2.Extensions(),
		})

		return
	}

	// Manage non common error
	c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
}
